requirements.txt
aiohttp==3.11.18
aiogram==3.20.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
certifi==2025.11.12
frozenlist==1.8.0
idna==3.11
magic-filter==1.0.12
multidict==6.7.0
propcache==0.4.1
pydantic==2.11.10
pydantic_core==2.33.2
python-dotenv==1.2.1
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.22.0
